Project new: renaming and styling
